Move AppConfig to conventional file
